requirements.txt
astroid==2.3.3
boto3==1.14.26
botocore==1.17.26
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
defusedxml==0.7.0rc1
Django
django-filter==2.3.0
django-mssql-backend==2.8.1
django-pyodbc-azure
django-s3direct==1.1.8
django-storages==1.9.1
djangorestframework==3.11.0
docopt==0.6.2
docutils==0.15.2
idna==2.10
importlib-metadata==1.7.0
isort==4.3.21
jmespath==0.10.0
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
mypy==0.761
mypy-extensions==0.4.3
mysql
mysqlclient==2.0.1
oauthlib==3.1.0
pipreqs==0.4.10
pycparser==2.20
PyJWT==1.7.1
pylint==2.4.4
pyodbc==4.0.30
python-dateutil==2.8.1
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
typed-ast==1.4.1
typing-extensions==3.7.4.1
urllib3==1.25.9
wincertstore==0.2
wrapt==1.11.2
yarg==0.1.9
zipp==3.1.0
numpy
pillow
django-cors-headers
emojis
